requirements.txt
aiofiles==0.5.0
aioredis==1.3.1
aniso8601==7.0.0
async-exit-stack==1.0.1
async-generator==1.10
async-timeout==3.0.1
asyncpg==0.23.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
ConfigArgParse==1.4.1
databases==0.4.3
dnspython==2.1.0
email-validator==1.1.2
fastapi==0.65.1
Flask==1.1.2
Flask-BasicAuth==0.2.0
gevent==21.1.2
geventhttpclient==1.4.4
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
h11==0.12.0
hiredis==2.0.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
locust==1.5.3
MarkupSafe==2.0.1
msgpack==1.0.2
orjson==3.5.2
passlib==1.7.4
promise==2.3
psutil==5.8.0
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-decouple==3.4
python-dotenv==0.17.1
python-multipart==0.0.5
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.25.1
Rx==1.6.1
six==1.16.0
SQLAlchemy==1.3.24
starlette==0.14.2
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
uvicorn==0.13.4
watchgod==0.7
websockets==8.1
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
